chore(deps): update JS/TS dependencies to latest (#441)
* chore(deps): update JS/TS dependencies to latest

- Bump all outdated root deps (prod + dev) and server deps to latest; pnpm-lock.yaml + server/package-lock.yaml regenerated.
- Hold typescript at 6.0.3: TS 7.0.x ships no stable JS compiler API (native port; API lands in 7.1) and typescript-eslint 8.64.0 declares peer typescript '>=4.8.4 <6.1.0'.
- eslint.config.mjs: unicorn/no-array-for-each renamed to unicorn/no-for-each (plugin v66 rename, same severity); jsx-a11y-x 0.2 configs surface (flatConfigs removed); @eslint-react/static-components turned off (new v5 preset rule, false positive on registry-lookup components — matches v4 behavior).
- Remove 300+ unnecessary type assertions flagged by typescript-eslint 8.59's hardened no-unnecessary-type-assertion; vi.hoisted mock widenings re-expressed as type annotations (no assertions).
- Reformat 43 files for prettier 3.9 style drift (prettier/prettier is error-level via eslint-plugin-prettier).
- Fix pre-existing failures: import-x/order + unused imports in 2 specs; check-dependency-boundaries.spec.ts asserted new RegExp() over an any-of path array (never matched) and contradicted the react rule's documented jsx-runtime coverage — now exercises the array with some() semantics.

* fix(commandinterface): drop unnecessary assertion after main merge

+++ src/modules/BrowserAi/workers/onnxInferenceWorker.ts
@@ -489,7 +489,7 @@ self.onmessage = async (event: MessageEvent<WorkerRequest>): Promise<void> => {
                 } else if (val.data instanceof Uint8Array) {
                     dataType = 'bool';
                 }
-                resultOutputs[key] = { data: val.data as Float32Array, dims: val.dims, type: dataType };
+                resultOutputs[key] = { data: val.data, dims: val.dims, type: dataType };
             }
             const response: WorkerResponse = {
                 type: 'inference-result',
